增加系统版本接口 增加根目录接口

+++ src/main/java/com/orbithy/cms/data/po/Classes.java
@@ -23,7 +23,7 @@ public class Classes {
 
     private String name;
     private String category;
-    private Integer point;
+    private float point;
     private Integer teacherId;
     private String classroom;
     private Integer weekStart;
